Moved Players Tables Out of Beta.
